move .hpp headers to .h
